Restore the schema back to its original state in case of failure
Usually, it takes a few rounds for the migration to work, which means the
fact that the 1.8 schema is left in 'public' makes 'setup.pl' think our
database has been migrated (but is left empty).

Since we need to retry: get the database back into a state which will let us
retry (and which will show the user his original data right where he left it).