Split process_and_run_upgrade_script to provide more entrypoints
In order to allow user-intervention during the upgrade, we need points
at which the server-process can be restarted in case a failure required
user-input.

This change splits process_and_run_upgrade_script to provide an additional
entrypoint to be called when schema upgrade checks fail during the
post-migration-script schema upgrade (from "the tag" to "the latest").

Note that we can't use "rebuild_modules" as the entrypoint, because it
has different "next steps" than those required to complete a migration.