Create traceable steps in upgrade workflows
Until today, the steps in the upgrade were hard-coded through a combination
of action parameters of buttons (in templates) and calling "next step
subroutines" in entrypoint routines after processing completed succesfully.

This commit completely reorganizes that: With a central dispatch table,
each workflow step is expected to dispatch through the central table,
keyed on its own step name.

This scheme decouples the flow from the steps and the templates,
which provides options to create more diversity for various types of
migrations.