Fix error thrown when opening the business unit 'edit' screen
When clicking on a business unit to view its details and edit it,
an error is thrown indicating that no locale has been set. Because
a locale is *always* initialized (if the user doesn't have one, 'en'
is selected) this error is an indication of a different problem.

The problem here is that the $business_unit shouldn't pose as a
$request -- ever.