Fix 'uninitialized variable' warnings in Log4perl middleware for old code
Note that the current implementation assumes the DynamicWorkflow middleware
preceeds it in its call chain. However, for old code, this isn't the case.
Instead, don't derive the name of the script from the call context, but from
the parameters passed on initialization. The action name can be had from the
request object.

Fixes #4606