Fix #4615: Reporting units from invoice don't filter Income statement
IS.pm and IR.pm failed to recognise and insert the reporting units from the
UI as values linked to the `acc_trans` table. (They succesfully link to the
`invoice` table, to reproduce the UI, though.)

Note that the `row_num` was already set up correctly in IS.pm; possibly, it
is part of an initial attempt to fix this problem. For IR.pm the same was
missing. This commit simply *uses* the row numbers to feed the INSERT.