Fix #4651: Tax on invoice off from search results
Note that the amounts posted in the books are correct and that
the search results show the amounts from the books (the correct ones).

However, when the invoice is presented, the amounts are recalculated
and the recalculation doesn't take the correct calculation precision into
account which results in the off-by-cents amounts.

This commit fixes the off-by-cents amounts by setting the rounding to
what was used when rendering the rows themselves, leading to correct
recalculation of the line totals (and thereby the correct tax amounts).
To make this explicit: it is the line total re-calculation which is off
due to the failed rounding. The tax is then calculated on the incorrectly
recalculated line total and by consequence incorrect as well, even though
there is nothing wrong with the tax calculation itself.

Whew.