Implement test infrastructure for COGS SQL routine testing through BDD
BDD is a mechanism which leads to much faster development of tests as well
as to much easier verification of the recipe in functional terms.

This commit introduces test infrastructure which runs BDD tests which aren't
tied to Selenium web-browser tests: we can drive the COGS SQL routines through
the DBI connection handle directly, eliminating the need for (slow) Selenium
access.