Remove dependency on dojo/domReady!
The only thing domReady does, is compensate for behaviour in IE which
exposes the 'onload' instead of the 'load' event -- but this is no longer
relevant, since we don't support IE anyway.

As a result, we are now switching to directly using the 'window Event' api
with the 'load' event, in preparation of a move to webpack which doesn't
deal with domReady.

Note that this commit also removes domReady from lsmb/Reconciliation as it
can't be required anymore, since that module will be loaded once the dom
has been ready for ages: it's typically loaded as part of loading an HTML
fragment being loaded in response to a menu-click into a DOM that's ready.