Fix save-after-update on assemblies
When saving assemblies, the 'currencies' value gets initialized with
a string value. Instead of depending on the value to be an array (or
uninitialized), simply overwrite the value with the array we need to
store in it.

This error isn't on master or 1.8 because the code was already changed
to what's in this commit.