Fix #4781: Cannot save batch in Payments due to date format
The existing code passes all values from the request straight into the
inner layers of the application, forfeiting its role as decoupling point
between the web and the inner operation of the application.

This commit separates the concerns of 'handling of web input' from
'marshalling payments into the database' into their respective
modules. This also means that the workflow script filters out the
contacts and invoices which have been indicated /not/ to be part of
the payment(s).