Change 'tax' to a tag with 'rate' sub-tags (and reduce account attributes)
No longer include 'contra' and 'obsolete' attributes when the value
indicates the default (false).

To indicate an account to be a tax account, a 'tax' tag must be included
as a subtag. The actual rates will be the subtags of the 'tax' tag. Without
rates, the 'tax' tag remains empty (but remains to indicate 'tax-ness').