Align templates between demo & demo_with_images
Where demo and xedemo are fundamentally different, where the LaTeX sources
are concerned, due to the fundamental differences between PDFLaTeX and
XeLaTeX, there is no such difference between demo and demo_with_images.

Conclusion: the templates should be the same, except for the code
which uses the images. This commit makes that so.