Backport syntax fixes from #4819 (templates cleanup)
Some of the PNL and B/S templates have incorrect syntax. Especially now
that these are moving to the database, it's important to have them be
correct: more than average attention is going to go their way.

Manual backport from a bigger change which itself shouldn't land verbatim
on the 1.8 branch.