Fix #3236: part description dropdown on invoice active after part selected
After a part has been selected, the drop-down on the part number disappears
because no further selection is required. The part description needs to stay
active, because the description can be edited. However, the drop-down also
remained active, which suggests a new part can be selected. This isn't the
case and this commit removes the dropdown showing up after the part has been
selected.