Fix #4949: Fix sequence number generator returning the name of the sequence
The existing code no longer works due to hash randomization (the fact that
it used to work is merely coincidental): due to hash randomization,
sometimes the 'value' key's value is returned as the first element and
some other times, the name of the sequence is returned.
The latter is what @aungzwin10 observes in the issue he reports.