Fix being unable to open employee+user screen
When there are two companies with sufficiently similar names,
(e.g. one's name being a subset of the other) opening the Contact
screen for an employee who is also a user, fails with an error
about the 'role_list' attribute of the user not matching the
specification. This is due to the fact that the query returns
NULL values in the result.