Fix #4872: Inability to sort AR > Search results for filter with spaces
Because the URL was round-tripped without URL-unescaping before passing
the URL back into the template again, the '%'-sign was re-escaped as
'%25' over and over. This also makes the filter behave different between
the first and second time the data is loaded. By unescaping between iterations,
the URL in the actual HTML page remains the same between iterations.

Manual backport from 1.9/master to prevent adding a new module dependency.