Close #3285: Session page accumulates endless numbers of sessions
Since there is no way to recover a session for which a browser tab
was closed/lost, there is no way to prevent endless growth to the number
of active sessions in the sessions table. The only remedy is to
make sessions always expire, even if it is after huge numbers of days.