Close #3286: Refactor parts search page
The radio buttons mentioned in the issue are now replaced with drop downs
(to save space and mouse movements). Hardly ever used 'parts details' search
parameters are now hidden and can be clicked open.

The entire thing is now much more visually appealing due to the use of
css with fieldsets and better element spacing.