Close #5113: Add selectable empty value to non-required SELECTs
The empty string is used by Dojo to indicate a separator line in the
drop-down. So, we can't use the empty string as an 'empty' indicator.
To work around that, we now use the special value "_!lsmb!empty!_"
to model the empty value -- which we filter out on the Perl side
to indicate that no value was selected.