Replace Try::Tiny with Syntax::Keyword::Try
Note that Syntax::Keyword::Try is supported as of Perl 5.14 and that
it has the desirable property that it allows using 'return' in the
'try', 'catch' and 'finally' blocks to return from the sub it's
used in (which Try::Tiny doesn't allow). Downside of it is that
Try::Tiny can be used in expressions out of the box, but this module
can not (needs experimental behaviour to be enabled to do so).