Initialize Log::Any to point to Log4perl
As PGObject::Util::DBAdmin is about to support Log::Any in its 1.4.0
release, initialize it to log to Log4perl which we use in our application.

Now that Log::Any has made its way into our code base (and now that I
understand its purpose), possibly, some of the other libraries published
by the project (or by me) on CPAN will move to it as well.