Move configuration of CoA and GIFI from SQL to structured data (XML)
Moving to structured data adds the benefit of consistency checking using
the XML schema verification.
Additionally, in a later iteration, it would be possible to parse (part)
of the content of the configuration file to extract a description which
could be presented to the user.

Also, since GIFI codes are in the CoA files (they exist in the account
creation SQL), there's a fixed combination between GIFI and CoA, so,
don't offer the choice anymore: GIFI are now incorporated into the
'setup XML'.