Fix failure to delete unapproved reconcilation report
In 1.8, the explicit reconciliations were added, adding a new
table to the schema. However, the removal of (unapproved)
recon reports was not expanded to remove the lines in the new
table. This approach is a more general approach: when report
lines are deleted, so are the items in the cross linking table.