Remove all uses of Company_Config other than PGNumber
As part of the quest to remove as much global state as possible, remove
as much use of Company_Config as possible: there's no need to have this
configuration in global state.

The remaining issue is with PGNumber, which depends on it for the
number of decimal places in number formatting. Replacing it entirely
seems rather straight forward, except for the uses in Report and
Report/Hierarchical which currently don't seem to have access to the
request object.