Replace iframe-based with Blob downloading
Downloading using Blobs and XHR2 is available now that we don't support
IE11 anymore. Blob-based downloading also provides us access to completion
events, which always was a problem with the iframe.

Additionally, errors ended up in the (invisible) iframe with the user
sitting and waiting for the response. With this approach error reporting
works through the regular error dialog.

Because Dojo doesn't provide access to the headers in the XHR response
(docs say it does, but really, it doesn't...), using xhr directly is
the best alternative.