Track history of Quotes/Orders/Invoices using workflows
Every quote, order or invoice gets an associated workflow. The actions in the
workflows only record history for now, since the actual code (old code) can't
yet be moved into the action implementations.
In order to be able to create a workflow for invoices, save the invoice
immediately upon creation. Due to the fact that the invoice is now saved when
created, this closes #5187 (order closed before invoice posted or saved).

In order to be able to save transactions without a transaction date,  this commit
relaxes the requirement to the 'transdate' field. Instead of requiring it to be
NOT NULL, it is required to be NOT NULL when the transaction is approved.