Fix #5289: Failure to save purchase shipping addresses (and contacts)
The code to save shipping addresses and contacts had only been implemented
on the sales side of things. By moving the code from IS to IIAA and updating
it to take both customer_id and vendor_id into account, the same functionality
now applies to both sales and purchase documents.