Fix inability to open saved invoices from search when gapless enabled
The problem is that the only thing that currently can reach the invoice
is the link that's being clicked which is the invoice number. However,
with gapless numbering, saved invoices don't get an invoice number, meaning
that no link gets rendered in their place...

Resolve the above mentioned issue by turning the ID column into a link to
the invoice as well -- and display the ID by default too.