Save any files that may be attached to deleted transactions
Before deleting the record from the `transactions` table (for which there
are no records in `ar`, `ap` or `gl`), we need to make sure not to
accidentally remove valuable files. Although these files are not accessible
through the UI and therefor effectively lost to the user, we should offer
the admin a possibility to recover by saving them on the server before
deleting forever.