Improve upgrade script based on real-life data
As we found real-life data where the records in the `transactions`
table were still linked to the `invoice` table and `invoice_tax_form`,
remove those records before deleting the ones in `transactions`.

Also change the namespace to which the salvaged files are stored,
in order to prevent collision of multiple databases (as much as
possible).