Add trial balance start date to MC trial balance verification
The table without the start date as part of the primary key failed on duplicate
rows when the last transaction in the database was part of the last snapshot.
(I.e. they had the same 'transdate'.) The duplicate comes from the fact that
next to all rows from the snapshots table, also a "start to end" trial balance
is added to the verification data. By adding the start date to the equation,
the two trial balances that end on the same date, are given a separate key,
because they start on different start dates (i.e. one at the beginning of all
time; the other at the end of the prior snapshot.