Prevent a duplicate key error during MC migration (edge-case)
When there's exactly *one* account snapshot *and* that snapshot includes
the last transaction date in the database, the trial balance interval
from the snapshots and the 'begin-to-end check' result in the same dates
which causes a duplicate key error.
By subtracting a day from the start-date of the 'begin-to-end check',
that check generates a primary key different from the account snapshots
check.