Re #2306: Enable filtering in customer/vendor dropdown (#5508)
In the AR/AP invoices and transactions, enable filtering on customers
and vendors (instead of providing a fixed list of selectable options).
Also enable this selection method by default for up to 9999 customers
or vendors. (Above that number, fall back to the current default which
means providing a text entry box to be used for seaching customers/vendors).

This does not back the filtering select with a service and hence does
not fully close the issue.