In the Income Statement, use the closing balance of the end date
In line with the option to select opening or closing balances in the
balance sheet report, remove the "ignore year-ends" option from
the income statement: it's confusing and doesn't relate to terminology
familiar to our type of end-user.
With the current settings, the PNL agrees with the default settings of
the balance sheet: 'Closing balance' shows the current earnings before
closing the books. So does the PNL report after this change.