Fix #5990: Incorrect indentation in hierarchical PNL and B/S reports
Apparently, using a single cell with "just" a colspan and a width, doesn't
work as expected. Instead, generate a series of cells, each with the
unit of indentation that's expected. This has the additional benefit that
the CSS only needs to list a single indentation level, because there's
going to be a series of repeated 1-level indentations.