Remove version comparisons from SL30 migration
The version comparison was a difference between 'current' and 1.6; the latter being
EOL these days, meaning that the version comparison has a constant (negative) outcome.

There's another reason why version comparisons are a bad idea: the migration target
schema for the migration should be constant. That is, it should be a specific version
of the schema determined by a tag in the 'sql/changes/LOADORDER' file. This tag returns
independent of the exact current LedgerSMB version, a constant database schema version.