Reconciliation fix (#6122)
Summarizing:

* Use named arguments instead of positional
Clean up for better readability

* Fix missing FROM clause in reconciliation
Fix using our_value as column name, our_value doesn't exist in cr_report_line

* Fix payment source doesn't correctly save to scn of cr_report_line

* Fix description always empty in reconciliation
entry_id from acc_trans is not saved in the cr_report_line table. It is saved in cr_report_line_links.

* Fix ID duplication on the payee view because of the cr_report_line_links

* Fix missing FROM caluse in sl3.0

* Undo change adding payments with the wrong granularity to the recon

And add a lot of documentation to (hopefully) explain why this code
looks the way it does.

* Document reconciliation test cases and re-order for clarity

Group the journal lines and add comments for clarity on the actual tests
that are intended to be executed.

* Add and adapt tests

Add test to check adjustment GL lines being added to payments given the same source and date
Adjust tests to incorporate Source as a splitting criterion *within* a payment
Adjust number of lines in test against account -11112 better distinguish test cases

Co-authored-by: Erik Huelsmann <ehuels@gmail.com>