Date entry: fix removing separator characters
Without this fix, the DateTextBox doesn't allow removal of content
beyond the first separator character (adding it back immediately on
removal), unless *all* text is selected and removed and entry is
started from scratch again.
With this fix, separators are only added when the input value is
appended/added to.