Implement the login page as a Vue application
This commit foregoes the implementation of the error dialog the old
implementation had; instead, it uses 'alert()' to pop up the error.

It solves the inputs on the front page being rendered multiple times
when mixing custom elements -- apparently, there are adverse effects
while Vue builds its virtual/shadow DOM.

In order to keep Vue's data in sync with the custom elements (which
cannot be done with the `v-model` directive, unfortunately), this
commit introduces the `v-update:<data-element>=""` directive.

On the custom elements implemented so far, there was no syncing
of values going on between the embedded widgets and the element
itself; this commit also implements the required mechanics for that.