Manually trigger Dojo's parser for pages that need it
This makes all pages go through the same initialization phase, dealing
with page specifics in consecutive `if` statements for clarity.

This also deals with explicitly hiding the "loading" div, if it's on the
page, which was previously hidden in "lsmb/main" (and that was its main
reason for it to still exist).