Fix deletion of *all* payment links
When deleting a batch, the procedure would delete *all* payment links;
after this change, there's now a restiction on the subselect relating
the DELETE to the subquery, restricting the results in the subquery.

(Before, *any* existance would delete *all* payment links, not just the
ones relating to the batch being deleted.)

While at it, remove code duplication.

Fixes #6368.