Fix double-counting of transactions in the trial balance
Before this commit, transaction lines with more than one
business unit associated are be double-counted in the balance
total due to the LEFT JOIN, which repeats the `acc_trans` line
for each business unit.

Instead, move the join to an EXISTS condition in the WHERE
clause where it's not part of the specification of the result
set.

Fixes 6402