Fix line descriptions not shown in AP transaction prints
There's overlap in naming the fields: there's a generic string field
`description`, which is invoice-level data. Then, there's an array
field called `description` which contains the descriptions of the rows.

For some reason this naming clash doesn't show when printing the AR
transactions, but it does show when printing AP transactions.

For the purpose of backporting, *add* a field called `linedescription`
which holds the same array of line-related descriptions. For 1.10,
we'll remove support for `description` holding the line descriptions,
but for a patch release, that's would lead to unacceptable breakage.

At the same time, update the templates so that new databases will be
created with future proof templates.

Fixes #6482.