Replace as_array with array_agg
There's a problem with the function signature of 'as_array' (which
is itself and aggregate fully implemented on top of `array_append`)
in Pg14, which is incompatible with our aggregate definition as well
as with the defition of `array_append` on prior versions of Pg.

The `array_agg()` function was introduced back in Pg8.4, well within
the bounds of our supported set of PostgreSQL versions.