Replace most `compound_array()` invocations by `array_agg()`
Most cases of `compound_array()` could have been written as calls
to `as_array()`, which was replaced by calls to the Pg native
unction `array_agg()`.

Reason to want removal of these aggregates is because of changes
in Pg14 which make pre-14 and 14+ versions of `array_agg` and
`array_cat` incompatible (and therefor require version-specific
definitions of the user-defined aggregates in LedgerSMB -- instead
I'd rather go an extra mile and try to remove the aggregates
alltogether).