Fix SQL error in reconciliation
When multiple payments have the same 'scn' (source) value *and* there's
at least one GL line with the same source value, the code wants to associate
that single line with both payments as a correction. That leads to an SQL error
('multiple rows returned in a query used as a value').

The solution is to exclude reconciliation lines where other lines with the
same 'scn' value exist, when considering GL lines a potential corrections.