Add 'Return' link to e-mail UI
When an e-mail workflow is in the 'CANCELLED' state, there's no
'Cancel' button, meaning there's no way to return from it, even
when there *is* a 'callback' parameter.
Additionally, a user may want to return using the callback, without
cancelling the e-mail workow.

This change adds a "<< Return" link to the e-mail UI when there's
a 'callback' parameter which enables both scenarios.