Prevent dismissable Toasts from displaying too short
Vue itself uses a 200ms delay for showing the LoaderComponent in the
router when loading takes too long. This implements a similar idea
for the Toasts, using a 250ms delay, because local testing shows that
there is a clear break with data taking either less than 200ms or far
more to load. 250ms then allows for a little latency and assures the
toast shows long enough.