Replace Hash::Merge config merging
Because Hash::Merge does not have a concept of "replace array
items" versus "append array items", it's impossible to specify
fewer array configuration items than the base configuration.

Being unable to override the base array means that it's
impossible to remove output formatters from the base setup,
which is an impractical limitation.