Eliminate `$request` as _render() argument
Instead of creating custom template variables with the need to pass
lots of ad-hoc information into the report class, have the `renderer()`
callbacks inject a `SETTINGS` template variable based on the company
settings.
Also eliminates the need to allocate a LedgerSMB::Setting object and
issue additional queries to find the relevant settings.