Separate 'save' and 'submit' concerns in recon
On one hand, this is simply 'better design'; on the other,
it's actually a requirement to be able to develop (BDD) test code
which is half-way sane (comprehensible).

In order to propagate state changes to the 'cr_report_line_links'
table, triggers are installed on the 'cr_report' and 'cr_report_line'
tables.