Fix transactions being saved without transaction date
In the 'ar' table, transactions have a default transaction date of
"now" (which is a timestamp value; better would be to use "today"
which is a date -- the same type as transdate, which is a date too).

However, when committing data to the 'acc_trans' table, the default
value for transdate in the 'ar' table, isn't taken into account,
leading to rows without 'transdate' values. When these rows are then
marked 'approved', the data integrity check kicks in and blocks the
transaction.