Remove LedgerSMB::DBObject::TransTemplate
This commit disables AR and AP template transactions. This looks
wrong, but there are so many issues with transaction templates that
I don't know where to start to describe them here. The short story
is that disabling them is a good thing; if there were no users for
the GL version, that should be disabled too, until there's a better
implementation.