Be more strict to allow non-uniquely reconciled lines during migration
Before this change, all lines that are assigned to multiple
reconciliations are marked 'unique_exempt'. However, this also
applies to lines which are part of reconciliations that still
need to be approved. We can - and need to be - more strict by
disallowing unapproved reconciliations to have lines overlapping
with approved reconciliations.

Note that this all applies to *the time of migration* __only__,
which is a relatively restricted interval.

Fixes #7078