Asset sale should take cash into account
This commit adds a Cash Account to asset disposal reports in order to
account for the cash component of the sale. Please do note that if an
invoice was created for the sale, there's some additional accounting
to be done to move the income out of the PNL, into the "cash" account.

There's a bigger plan for rewriting fixed assets support. This commit
is part of the "old world", not the start of development of the new,
shiny and highly desired code base.

Fixes #5180.