Follow-up to #7048: Improve old-UI integration with Vue
With the widgets destroyed in `beforeUpdate`, it turns out that we
can still get stuck with a UI "dead in the water": when the UI update
fails with an error, the widgets have already been destroyed, but the
UI is still visible (and looks like it can be used after closing the
error dialog). By destroying the widgets just before they get replaced,
the opportunities for this scenario are much reduced (if not eliminated).