Don't create a new part when no parts match
Note that the "no parts match" condition *only* occurs when the
system holds *no* other parts, because the UI will fill in the
first part's number when there *are* parts defined.

To prevent setting the expectation that this is how parts are
created, don't do it for the first one either.

Closes #6530