Show reversing reference when reversal is posted
When a reversing transaction is created the GUI shows the reference and
ID of the reversed transaction. If the user presses "save" only the ID
is shown. When the user presses "save" the only reference LedgerSMB has
to the reversed transaction is the transaction ID from the database. To
show the reference it needs to be collected.

The reason why the reference show in the first step, but not in the
second, seems to be two-fold:

1. The reference data is not passed to the next view when pressing
   "save".
2. The form data is reset once in the post subroutine (which is called
   on pressing "save") so the data doesn't reach the display subroutine.

When the user press "reverse" the form data needed in the next view is
passed as hidden form fields in the HTML. To preserve other form data it
is scoped to be local to a do{...} block. Without this it would be
overwritten with data from the database and since the references to the
reversing has not been saved the unscoped fields will be reset or
undefined. As this happens before rendering the form the displayed form
will not have the expected data.