Show reversing reference for AR reversals
When reversing an AR transaction the onscreen hint does not show the
reference number, i.e. the invoice number, of the transaction that is
being reversed. This prompt is shown when the user presses the "revert"
button in the Accounts Receivable (AR) section, when viewing a
transaction. What the user sees:

"This transaction will reverse transaction Y"

What the user should see:

"This transaction will reverse transaction X with ID Y"

Where Y is the database ID of the transaction and X is the invoice
number (i.e. reference number) of the transaction.

This reference/invoice number is not set as a field on the form, during
the reverse subroutine and thus not available to the following
subroutines that displays the hint. A reference to this number is also
missing in the string template, which makes up the hint, so even if
available it would not show.