Show reversing reference for DI reversals
When reversing a Debit Invoice (DI) or Vendor Invoice (VI the onscreen
hint does not show the reference number, i.e. the invoice number, of the
invoice that is being reversed. This prompt is shown when the user
presses the "reverse" button in the Accounts Payable (AP) section, when
viewing a Debit/Vendor Invoice. What the user sees:

"This transaction will reverse transaction Y"

What the user should see:

"This transaction will reverse transaction X with ID Y"

Where Y is the database ID of the transaction and X is the invoice
number (i.e. reference number) of the invoice.

This reference/invoice number is not set as a field on the form, during
the reverse subroutine and thus not available to the following
subroutines that displays the hint. A reference to this number is also
missing in the string template, which makes up the hint, so even if
available it would not show.