Increase max-num comparison periods for reports (#7468) (#7469)
When creating reports like Income Statement and Balance Sheet it is
useful to be able to split the year into months to be able to compare a
whole year. It may also be interesting to compare with the last month of
the last year. To accommodate this usecase we increase the maximum
number of comparison periods from 9 to 13.