Fix encoding of the column data, including col_id
Due to the fact that all column data was encoded for output-format-safety,
the CoA and Trial Balance reports didn't show some columns in PDF format
(with LaTeX intermediate representation). The column *keys* in the ROW data
do *not* get encoded, whereas the column keys (stored in the column data as
values) *do* get encoded. Due to this mismatch, the column name 'credit_balance'
differed between the row data and the column data (the latter containing the
encoded value 'credit\_balance').

Fixes #7739.