Add visual feedback of payment line being unapproved
Payment lines which are part of a payment batch may be
unapproved even if the transaction itself *is* approved.

The fact that nothing indicates a line may be unapproved,
means that the user will be in confusion as to why the
transaction isn't marked (partially) 'paid' in the
outstanding report.