Fix generated HTML to include closing 'html' tag
By making the header into a wrapper, appending the html closing
tag (and other epilogue stuff) is ensured to happen.

The old code expected to add a closing tag, but really didn't,
because '[% end_html %]' is a variable insertion command,
not a block execution... Resulting in broken HTML being
returned.