Fix creating an invoice from an order
The 'period-closed' condition returns "true" when 'transdate' is missing,
meaning that '!period-closed' returns "false" thus preventing the "post"
(="save") action from being triggered when no 'transdate' has been set.

However: 'transdate' is explicitly empty when creating an invoice from
an order to force the UI to apply its default vaule of "today" (instead
of using the 'transdate' of the order, which is the order issue date).

Fixes #7855