Remove 'entity_class' from entity table
The reasoning is that an entity has multiple roles,
where the customer and vendor roles are modeled using
'entity_credit_account', employees using the equally
named table, etc.

Whether an entity is a company, person or 'robot' depends
on whether a join can be made with any of the three and
isn't modelled by the entity_class anyway.