Decouple the E-mail workflow from the persister
By having persister-behaviour associated with specific context values,
actions can stick to mutating the content (instead of modifying the database)
and the persister can modify the database.

By introducing tied values the persister can delay reading values from the
database until they are requested by the workflow -- all seamlessly.  Which
is a requirement for processing of the file content of attachments, to save
memory use/bloat.