Fix cogs calculation for assemblies
The root cause of all this is the fact that the sellprice of the assemblies
is calculated and stored as a negative number.  To make things work going
forward, all we had to do was to make the number positive.  However, to fix
all the historic data, a largish migration script was required which in turn
required a new batch and transaction type because none of the existing ones
looked particularly useful for this purpose.

Fixes #8351