Remove myconfig{acs} references
The ACS thingy is from a distant past. These days we rely on the database
to provide the correct access (and to prevent it).  The 'acs' key hasn't
been populated in %myconfig for a loooong time, making conditional code
blocks dead code, except for negative conditions, which are part of the
main line of code.