Reorganize ShipTo screen for better alignment
Put the "new location" and "new contact" entry rows in the
same table as the listing of the existing ones: this aligns
the column titles with the inputs, making the screen more
self-explanatory.

At the same time, fix the "contacts" section to have radio
buttons to select the appropriate contact information.  Do
note that there's no way to store the selected value; the
best way forward is to be determined.