feat: add '--prepare-only' option to 'create' command
This commit adds a '--prepare-only' option to the 'ledgersmb-admin create' command, allowing users to prepare an existing database without attempting to create it. This is particularly useful in deployment scenarios where the database is pre-created, such as when using managed database services like DigitalOcean Managed Databases.

The '--prepare-only' option:

- Skips the database creation step, avoiding errors when the database already exists.
- Performs the preparation steps by loading the base schema, applying changes, and loading modules into the specified database.
- Provides a consistent user experience by integrating the functionality into the existing 'create' command.

This addition addresses the need to prepare existing databases in situations where the 'create' command cannot be used because it attempts to create the database, which may not be possible or desired in certain environments.

This is inspired from the workaround for issue:  #8411