Move logout to its own workflow script
'login.pl' discards existing sessions and generates a new one, making the existing
one inaccessible for logging out. 'erp.pl' requires a valid connection to the database,
meaning that it requires a valid connection, but we want 'logout' to work when the
session is expired (that is: discard the error that the session is expired and report
success).
To achieve that, the 'erp.pl', 'login.pl' and new 'logout.pl' workflow scripts each
have their own middleware configuration.