Update UI only from last click
This prevents any in-flight page loads updating the UI when a newer page
has already been loaded. This can happen when loading one menu item takes
long and the user proceeded to clicking another menu item.

This also prevents unloading the page when it's in a state that isn't
unloadable, because Dojo's parser is working on it.